# src/main/resources/api/openapi.yaml
openapi: 3.0.0
info:
  title: Conduit
  version: '1.0'
  description: 'The Conduit Project API is a RESTful service that allows developers to interact with a social blogging platform. The API provides endpoints for managing users, articles, comments, profiles, and tags. It supports common actions such as authentication, article creation, comment posting, and user profile management'
  contact:
    url: 'https://thinkster.io'
    name: Thinkster
    email: hello@thinkster.io
  termsOfService: 'https://raw.githubusercontent.com/gothinkster/realworld/main/LICENSE'
  license:
    url: 'https://raw.githubusercontent.com/gothinkster/realworld/main/LICENSE'
    name: MIT license
servers:
  - url: 'http://localhost:8080'
paths:
...
  /users:
    post:
      summary: ''
      operationId: users-post
      responses:
        '201':
          description: ''
          headers: {}
          content: {}
        '400':
          $ref: '#/components/responses/ResponseError'
        '401':
          $ref: '#/components/responses/ResponseError'
        '403':
          $ref: '#/components/responses/ResponseError'
        '404':
          $ref: '#/components/responses/ResponseError'
        '500':
          $ref: '#/components/responses/ResponseError'
      requestBody:
        $ref: '#/components/requestBodies/User'
      description: Create resource
      tags:
        - users
        - post
      parameters: []
      x-internal: false
    get:
      summary: ''
      operationId: users-list
      responses:
        '200':
          $ref: '#/components/responses/UserList'
        '400':
          $ref: '#/components/responses/ResponseError'
        '401':
          $ref: '#/components/responses/ResponseError'
        '403':
          $ref: '#/components/responses/ResponseError'
        '404':
          $ref: '#/components/responses/ResponseError'
        '500':
          $ref: '#/components/responses/ResponseError'
      description: List resources
      tags:
        - users
        - list
      parameters:
        - $ref: '#/components/parameters/offset'
        - $ref: '#/components/parameters/limit'
      security: []
    parameters: []
...
components:
  schemas:
    BaseEntity:
      type: object
      x-examples: {}
      properties:
        id:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - createdAt
        - updatedAt
    Pagination:
      title: Pagination
      x-stoplight:
        id: icmrnesjmetu3
      type: object
      x-examples: {}
      properties:
        offset:
          type: integer
          x-stoplight:
            id: 2av3litdqr1sc
          minimum: 0
          default: 0
        page:
          type: integer
          x-stoplight:
            id: 60kdo2v5i26ba
          default: 1
          minimum: 1
        limit:
          type: integer
          x-stoplight:
            id: 8wfqg4ashis34
          default: 10
          minimum: 10
          maximum: 100
        pages:
          type: integer
          x-stoplight:
            id: rl7hu7gmb4tm5
          default: 1
          minimum: 1
      required:
        - offset
        - page
        - limit
        - pages
    User:
      type: object
      x-examples: {}
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        username:
          type: string
          maxLength: 50
          minLength: 1
          description: Unique value
          pattern: '^[a-zA-Z][a-zA-Z0-9_]*$'
        email:
          type: string
          format: email
          description: Unique value
          minLength: 5
          maxLength: 100
        bio:
          type: string
          maxLength: 500
          nullable: true
        image:
          type: string
          format: uri
          nullable: true
      required:
        - name
        - username
        - email
...
    ResponseError:
      type: object
      x-stoplight:
        id: pu7ozc4hb0ns0
      x-examples: {}
      properties:
        status:
          type: integer
          enum:
            - 500
            - 404
            - 403
            - 401
            - 400
        message:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: email
              issue:
                type: string
                example: Email should be a email
            required:
              - field
              - issue
      required:
        - status
        - message
  responses:
...
    ResponseError:
      description: Response error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResponseError'
          examples:
            '400':
              $ref: '#/components/examples/ResponseError400'
            '401':
              $ref: '#/components/examples/ResponseError401'
            '403':
              $ref: '#/components/examples/ResponseError403'
            '404':
              $ref: '#/components/examples/ResponseError404'
            '500':
              $ref: '#/components/examples/ResponseError500'
    UserList:
      description: User list response
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/BaseEntity'
                    - $ref: '#/components/schemas/User'
              pagination:
                $ref: '#/components/schemas/Pagination'
            required:
              - data
              - pagination
          examples:
            Example 1:
              $ref: '#/components/examples/UserList'
...
  parameters:
...
    offset:
      name: offset
      in: query
      schema:
        type: number
        minimum: 0
      description: Pagination offset
    limit:
      name: limit
      in: query
      required: false
      schema:
        type: number
        minimum: 1
        maximum: 100
      description: Pagination limit
  examples:
    UserList:
      value:
        data:
          - id: 8e5e5579-71b9-b6e4-1981-25deb7193c58
            createdAt: '1991-01-24T00:00:00.0Z'
            updatedAt: '2007-08-13T23:00:00.0Z'
            name: consectetur
            username: OBLI
            email: U53WPN5veXZ@EevETTGnoEKslZu.pumt
            image: 'https://vjvlzYKaxKTrDAERUoMPGNckWQWpob.buXV9bc5iSSm+PGp46UJwVEUdRVCftC+OPTT+5Ok4hC-zqZRD4nYMt52hCs'
            bio: esse
        pagination:
          limit: 17
          offset: 58731375
          page: 74522413
          pages: 16885263
...
    ResponseError400:
      value:
        status: 400
        message: Bad Request
        details:
          - field: email
            issue: Email should be a email
    ResponseError500:
      value:
        status: 500
        message: Internal Server Error
    ResponseError401:
      value:
        status: 401
        message: Unauthorized
        details:
          - field: authorization
            issue: Token is missing or invalid
    ResponseError403:
      value:
        status: 403
        message: Forbidden
        details:
          - field: access
            issue: User does not have the necessary permissions
    ResponseError404:
      value:
        status: 404
        message: Not Found
  securitySchemes: {}
  requestBodies:
    User:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/User'
          examples: {}
...